Added support for new Huion Tilt functionality
The patch consists of:

  - Patch the report descriptor to add the tilt fields

  - Invert the Y tilt value in the raw events

The tilt data value range is between C4 and 3C (+-60).  it is located
on bytes 10 and 11 of the report as follows:

 00 01 02 03 04 05 06 07 08 09 10 11
  ^  ^  ^  ^  ^  ^  ^  ^  ^  ^  ^  ^
  |  |  |  |  |  |  |  |  |  |  |  |
  |  |  |  |  |  |  |  |  |  |  |  Y Tilt
  |  |  |  |  |  |  |  |  |  |  X Tilt
  |  |  |  |  |  |  |  |  |  Y HH
  |  |  |  |  |  |  |  |  X HH
  |  |  |  |  |  |  |  Pressure H
  |  |  |  |  |  |  Pressure L
  |  |  |  |  |  Y H
  |  |  |  |  Y L
  |  |  |  X H
  |  |  X L
  |  Pen buttos
  Report ID - 0x08

After uclogic_raw_event it looks like this:

 00 01 02 03 04 05 06 07 08 09 10 11
  ^  ^  ^  ^  ^  ^  ^  ^  ^  ^  ^  ^
  |  |  |  |  |  |  |  |  |  |  |  |
  |  |  |  |  |  |  |  |  |  |  | -Y Tilt
  |  |  |  |  |  |  |  |  |  |  X Tilt
  |  |  |  |  |  |  |  |  |  Pressure H
  |  |  |  |  |  |  |  |  Pressure L
  |  |  |  |  |  |  |  Y HH
  |  |  |  |  |  |  Y H
  |  |  |  |  |  Y L
  |  |  |  |  X HH
  |  |  |  X H
  |  |  X L
  |  Pen buttos
  Report ID - 0x08 : From report decriptor REPORT_ID

Important:

This patch was only tested on a Huion H610Pro V2, my guess
is that it may mess up tablets without tilt, as I did
not add any logic to differentiate between for example,
the H610Pro and the H610Pro V2.